refactor(EWM-492): The data-layer structure has been transformed in accordance with the convention
