+++ lib/feature/presets_config/data/release_notes.dart
@@ -1,12 +1,17 @@
-import 'package:app/feature/presets_config/data/model/release_note.dart';
+import 'package:app/feature/presets_config/data/release_note.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'release_notes.freezed.dart';
 part 'release_notes.g.dart';
 
+/// Collection of release notes for different app versions.
+///
+/// Maps version identifiers to their corresponding release note information.
 @freezed
 class ReleaseNotes with _$ReleaseNotes {
   const factory ReleaseNotes({
+    /// Map of version identifiers to their corresponding release notes.
+    /// The key is typically a version string like "1.0.0".
     required Map<String, ReleaseNote> notes,
   }) = _ReleaseNotes;
