+++ lib/feature/presets_config/presets_config.dart
@@ -1,7 +1,7 @@
-export 'data/model/preset_config_type.dart';
-export 'data/model/connection_network_dto.dart';
-export 'data/model/release_note.dart';
-export 'data/model/release_notes.dart';
-export 'data/model/update_rules.dart';
+export 'data/preset_config_type.dart';
+export 'data/connection_network_dto.dart';
+export 'data/release_note.dart';
+export 'data/release_notes.dart';
+export 'data/update_rules.dart';
 export 'data/presets_config_exceptions.dart';
-export 'data/presets_config_reader.dart';
+export 'domain/presets_config_reader.dart';
